Split template-source runtime boundaries
